Created new index file and datastore directory
